remove jwt authentication due to some error

+++ src/main/java/com/example/demo/repositories/StudentRepository.java
@@ -3,9 +3,5 @@ package com.example.demo.repositories;
 import com.example.demo.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Set;
-
-public interface StudentRepository extends JpaRepository<Student, Integer> {
-
-       // Set<Student>findBySubjectId(Integer subjectId);
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
